Fixing Static Errors that were appearing when installing this module.
I was getting a static error that said "function extraStatics was deprecated and should just be set" so I have set "private static $db" instead of extrastatics

I was also getting an error with how the update CMS fields was being set, where it wasn't returning fields, but just trying to change it on the fly. This was also causing an error, so I set it as "Fieldlist $fields" and return $fields at the end of the function call.